[WIP] mis_builder refactoring: split mis_builder.py in two
mis_report.py has the heavy stuff, and contains all the computation
and rendering code
