+++ mis_builder/models/__init__.py
@@ -2,6 +2,7 @@
 # © 2014-2015 ACSONE SA/NV (<http://acsone.eu>)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import mis_builder
+from . import mis_report
+from . import mis_report_instance
 from . import mis_builder_style
 from . import aep
